:white_check_mark: add tests geojson, move tests infrastructure right file name
