+++ geotrek/infrastructure/tests/test_filters.py
@@ -1,6 +1,13 @@
+import datetime
+
+from django.test import TestCase
+
+from geotrek.authent.tests.base import AuthentFixturesTest
+from geotrek.authent.tests.factories import PathManagerFactory
 from geotrek.infrastructure.filters import InfrastructureFilterSet
-from geotrek.infrastructure.tests.factories import InfrastructureUsageDifficultyLevelFactory, InfrastructureFactory, InfrastructureMaintenanceDifficultyLevelFactory
-from django.test.testcases import TestCase
+from geotrek.infrastructure.tests.factories import (InfrastructureUsageDifficultyLevelFactory, InfrastructureFactory,
+                                                    InfrastructureMaintenanceDifficultyLevelFactory)
+from geotrek.maintenance.tests.factories import InterventionFactory
 
 
 class DifficultyLeversFilterTest(TestCase):
@@ -50,3 +57,99 @@ class DifficultyLeversFilterTest(TestCase):
         self.assertIn(self.infra_u1_m2, qs)
         self.assertNotIn(self.infra_u2_m3, qs)
         self.assertNotIn(self.infra_m2, qs)
+
+
+class InfraFilterTestMixin:
+    factory = None
+    filterset = None
+
+    def login(self):
+        user = PathManagerFactory(password='booh')
+        self.client.force_login(user=user)
+
+    def test_intervention_filter(self):
+        self.login()
+
+        model = self.factory._meta.model
+        # We will filter by this year
+        year = 2014
+        good_date_year = datetime.datetime(year=year, month=2, day=2)
+        bad_date_year = datetime.datetime(year=year + 2, month=2, day=2)
+
+        # Bad topology/infrastructure: No intervention
+        self.factory()
+
+        # Bad signage: intervention with wrong year
+        bad_topo = self.factory()
+        InterventionFactory(target=bad_topo, date=bad_date_year)
+
+        # Good signage: intervention with the good year
+        good_topo = self.factory()
+        InterventionFactory(target=good_topo, date=good_date_year)
+
+        data = {
+            'intervention_year': year
+        }
+        response = self.client.get(model.get_datatablelist_url(), data)
+
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual(len(response.json()['data']), 1)
+
+    def test_duplicate_implantation_year_filter(self):
+        self.login()
+
+        model = self.factory._meta.model
+        # We will check if this
+        year = 2014
+        year_t = datetime.datetime(year=year, month=2, day=2)
+
+        # Bad signage: intervention with wrong year
+        topo_1 = self.factory()
+        InterventionFactory(target=topo_1, date=year_t)
+
+        # Good signage: intervention with the good year
+        topo_2 = self.factory()
+        InterventionFactory(target=topo_2, date=year_t)
+
+        response = self.client.get(model.get_list_url())
+        self.assertContains(response, '<option value="2014">2014</option>', count=1)
+
+
+class InfrastructureFilterTest(InfraFilterTestMixin, AuthentFixturesTest):
+    factory = InfrastructureFactory
+    filterset = InfrastructureFilterSet
+
+    def test_none_implantation_year_filter(self):
+        self.login()
+        model = self.factory._meta.model
+        InfrastructureFactory.create()
+        response = self.client.get(model.get_list_url())
+        self.assertFalse('option value="" selected>None</option' in str(response))
+
+    def test_implantation_year_filter(self):
+        self.login()
+        model = self.factory._meta.model
+        i = InfrastructureFactory.create(implantation_year=2015)
+        i2 = InfrastructureFactory.create(implantation_year=2016)
+        response = self.client.get(model.get_list_url())
+
+        self.assertContains(response, '<option value="2015">2015</option>')
+        self.assertContains(response, '<option value="2016">2016</option>')
+
+        filter = InfrastructureFilterSet(data={'implantation_year': [2015]})
+        self.assertTrue(i in filter.qs)
+        self.assertFalse(i2 in filter.qs)
+
+    def test_implantation_year_filter_with_str(self):
+        filter = InfrastructureFilterSet(data={'implantation_year': 'toto'})
+        self.login()
+        model = self.factory._meta.model
+        i = InfrastructureFactory.create(implantation_year=2015)
+        i2 = InfrastructureFactory.create(implantation_year=2016)
+        response = self.client.get(model.get_list_url())
+
+        self.assertContains(response, '<option value="2015">2015</option>')
+        self.assertContains(response, '<option value="2016">2016</option>')
+
+        self.assertIn(i, filter.qs)
+        self.assertIn(i2, filter.qs)
