+++ geotrek/land/tests/test_views.py
@@ -71,10 +71,21 @@ class PhysicalEdgeViewsTest(CommonTest):
     model = PhysicalEdge
     modelfactory = PhysicalEdgeFactory
     userfactory = PathManagerFactory
-    get_expected_json_attrs = None  # Disable API tests
     extra_column_list = ['eid']
     expected_column_list_extra = ['id', 'physical_type', 'eid']
     expected_column_formatlist_extra = ['id', 'physical_type', 'eid']
+    expected_json_geom = {'coordinates': [[3.0, 46.5], [3.001304, 46.5009004]],
+                          'type': 'LineString'}
+
+    def get_expected_geojson_geom(self):
+        return self.expected_json_geom
+
+    def get_expected_geojson_attrs(self):
+        return {
+            'id': self.obj.pk,
+            'name': self.obj.name,
+            'color_index': self.obj.physical_type_id
+        }
 
     def get_good_data(self):
         path = PathFactory.create()
@@ -97,10 +108,22 @@ class LandEdgeViewsTest(CommonTest):
     model = LandEdge
     modelfactory = LandEdgeFactory
     userfactory = PathManagerFactory
-    get_expected_json_attrs = None  # Disable API tests
     extra_column_list = ['owner', 'agreement']
     expected_column_list_extra = ['id', 'land_type', 'owner', 'agreement']
     expected_column_formatlist_extra = ['id', 'owner', 'agreement']
+    expected_json_geom = {'coordinates': [[3.0013501, 46.5008686],
+                                          [3.0000461, 46.4999682]],
+                          'type': 'LineString'}
+
+    def get_expected_geojson_geom(self):
+        return self.expected_json_geom
+
+    def get_expected_geojson_attrs(self):
+        return {
+            'id': self.obj.pk,
+            'name': self.obj.name,
+            'color_index': self.obj.land_type_id
+        }
 
     def get_good_data(self):
         path = PathFactory.create()
@@ -123,10 +146,22 @@ class CompetenceEdgeViewsTest(CommonTest):
     model = CompetenceEdge
     modelfactory = CompetenceEdgeFactory
     userfactory = PathManagerFactory
-    get_expected_json_attrs = None  # Disable API tests
     extra_column_list = ['eid']
     expected_column_list_extra = ['id', 'organization', 'eid']
     expected_column_formatlist_extra = ['id', 'organization', 'eid']
+    expected_json_geom = {'coordinates': [[2.9999539, 46.5000318],
+                                          [3.0012579, 46.5009323]],
+                          'type': 'LineString'}
+
+    def get_expected_geojson_geom(self):
+        return self.expected_json_geom
+
+    def get_expected_geojson_attrs(self):
+        return {
+            'id': self.obj.pk,
+            'name': self.obj.name,
+            'color_index': self.obj.organization_id
+        }
 
     def get_good_data(self):
         path = PathFactory.create()
@@ -149,10 +184,22 @@ class WorkManagementEdgeViewsTest(CommonTest):
     model = WorkManagementEdge
     modelfactory = WorkManagementEdgeFactory
     userfactory = PathManagerFactory
-    get_expected_json_attrs = None  # Disable API tests
     extra_column_list = ['eid']
     expected_column_list_extra = ['id', 'organization', 'eid']
     expected_column_formatlist_extra = ['id', 'organization', 'eid']
+    expected_json_geom = {'coordinates': [[2.9999078, 46.5000637],
+                                          [3.0012118, 46.5009641]],
+                          'type': 'LineString'}
+
+    def get_expected_geojson_geom(self):
+        return self.expected_json_geom
+
+    def get_expected_geojson_attrs(self):
+        return {
+            'id': self.obj.pk,
+            'name': self.obj.name,
+            'color_index': self.obj.organization_id
+        }
 
     def get_good_data(self):
         path = PathFactory.create()
@@ -175,10 +222,22 @@ class SignageManagementEdgeViewsTest(CommonTest):
     model = SignageManagementEdge
     modelfactory = SignageManagementEdgeFactory
     userfactory = PathManagerFactory
-    get_expected_json_attrs = None  # Disable API tests
     extra_column_list = ['eid']
     expected_column_list_extra = ['id', 'organization', 'eid']
     expected_column_formatlist_extra = ['id', 'organization', 'eid']
+    expected_json_geom = {'coordinates': [[3.0013962, 46.5008368],
+                                          [3.0000922, 46.4999363]],
+                          'type': 'LineString'}
+
+    def get_expected_geojson_geom(self):
+        return self.expected_json_geom
+
+    def get_expected_geojson_attrs(self):
+        return {
+            'id': self.obj.pk,
+            'name': self.obj.name,
+            'color_index': self.obj.organization_id
+        }
 
     def get_good_data(self):
         path = PathFactory.create()
